requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-mistralai
chromadb
sentence-transformers
pymupdf
python-dotenv
pydantic
protobuf>=3.20
tqdm
torch
torchvision